Update README.md
The Perceptron algorithm is a fundamental neural network structure utilized for binary classification tasks. It consists of two layers: an input layer and an output layer, interconnected by weighted connections. The primary objective of this algorithm is to identify a linear decision boundary that can effectively separate two distinct classes. This is accomplished by iteratively adjusting the weights and bias term. The perceptron employs a signed linear function as its activation function, and the learning process involves updating the weights whenever the predicted label differs from the true label. In this particular project, I undertook the task of implementing the binary perceptron algorithm entirely from the ground up.